feat: Add comprehensive queue testing and fix FindFailedJob int/long bug (QDC-17)
Add 74 new tests covering EfQueueContextAdapter, cross-provider parity between
SQLite and EF providers, end-to-end dispatch lifecycle, QueueRunner lifecycle,
serialization edge cases, and IJobDispatcher interface compliance. Fix int→long
cast bug in EfQueueContextAdapter.FindFailedJob that caused ArgumentException.

+++ src/NoMercy.Queue.MediaServer/EfQueueContextAdapter.cs
@@ -134,7 +134,7 @@ public class EfQueueContextAdapter : IQueueContext
 
     public FailedJobModel? FindFailedJob(int id)
     {
-        FailedJob? entity = _context.FailedJobs.Find(id);
+        FailedJob? entity = _context.FailedJobs.Find((long)id);
         return entity == null ? null : ToFailedModel(entity);
     }
